Change the package of the metrics classes
