+++ component/src/main/java/io/siddhi/extension/execution/file/FileCopyExtension.java
@@ -31,7 +31,7 @@ import io.siddhi.core.query.processor.ProcessingMode;
 import io.siddhi.core.query.processor.stream.function.StreamFunctionProcessor;
 import io.siddhi.core.util.config.ConfigReader;
 import io.siddhi.core.util.snapshot.state.StateFactory;
-import io.siddhi.extension.io.file.util.metrics.FileCopyMetrics;
+import io.siddhi.extension.io.file.metrics.FileCopyMetrics;
 import io.siddhi.extension.util.Utils;
 import io.siddhi.query.api.definition.AbstractDefinition;
 import io.siddhi.query.api.definition.Attribute;
@@ -143,7 +143,10 @@ public class FileCopyExtension extends StreamFunctionProcessor {
         if (MetricsDataHolder.getInstance().getMetricService() != null &&
                 MetricsDataHolder.getInstance().getMetricManagementService().isEnabled()) {
             String siddhiAppName = siddhiQueryContext.getSiddhiAppContext().getName();
-            fileCopyMetrics = new FileCopyMetrics(siddhiAppName);
+            if (MetricsDataHolder.getInstance().getMetricManagementService().isReporterRunning(
+                    "prometheus")) {
+                fileCopyMetrics = new FileCopyMetrics(siddhiAppName);
+            }
         }
         return null;
     }
