+++ component/src/main/java/io/siddhi/extension/execution/file/FileMoveExtension.java
@@ -31,7 +31,7 @@ import io.siddhi.core.query.processor.ProcessingMode;
 import io.siddhi.core.query.processor.stream.function.StreamFunctionProcessor;
 import io.siddhi.core.util.config.ConfigReader;
 import io.siddhi.core.util.snapshot.state.StateFactory;
-import io.siddhi.extension.io.file.util.metrics.FileMoveMetrics;
+import io.siddhi.extension.io.file.metrics.FileMoveMetrics;
 import io.siddhi.extension.util.Utils;
 import io.siddhi.query.api.definition.AbstractDefinition;
 import io.siddhi.query.api.definition.Attribute;
@@ -141,7 +141,10 @@ public class FileMoveExtension extends StreamFunctionProcessor {
         if (MetricsDataHolder.getInstance().getMetricService() != null &&
                 MetricsDataHolder.getInstance().getMetricManagementService().isEnabled()) {
             String siddhiAppName = siddhiQueryContext.getSiddhiAppContext().getName();
-            fileMoveMetrics = new FileMoveMetrics(siddhiAppName);
+            if (MetricsDataHolder.getInstance().getMetricManagementService().isReporterRunning(
+                    "prometheus")) {
+                fileMoveMetrics = new FileMoveMetrics(siddhiAppName);
+            }
         }
         return null;
     }
